+++ app/src/main/java/com/shov/unlimstorage/views/navigations/SettingsNavigation.kt
@@ -5,8 +5,8 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.shov.unlimstorage.values.Screen
-import com.shov.unlimstorage.viewModels.singletonViewModel
-import com.shov.unlimstorage.viewModels.updateViewModel
+import com.shov.unlimstorage.viewModels.provider.singletonViewModel
+import com.shov.unlimstorage.viewModels.provider.updateViewModel
 import com.shov.unlimstorage.views.settings.SettingsScreen
 import com.shov.unlimstorage.views.settings.accounts.AccountsScreen
 import com.shov.unlimstorage.views.settings.updates.UpdateScreen
